+++ packages/manager/src/store/linodes/linodes.helpers.test.ts
@@ -16,8 +16,8 @@ const maintenanceNotification: (
       type: 'linode',
       url: 'https://hello.world'
     },
-    when: '2019-10-17 21:52:28',
-    until: '2019-10-17 21:52:28',
+    when: 'rightnow',
+    until: 'later',
     body: null
   }
 ];
@@ -30,29 +30,20 @@ describe('Linode Redux Helpers', () => {
       {
         ...linode1,
         maintenance: {
-          when: '2019-10-17 21:52:28',
-          until: '2019-10-17 21:52:28',
+          when: 'rightnow',
+          until: 'later',
           type: 'reboot'
         }
       }
     ]);
   });
 
-  it('returns Linode with `maintenance: null` if there are no notifications', () => {
-    expect(
-      addNotificationsToLinodes(maintenanceNotification(4325345345), [linode1])
-    ).toEqual([
-      {
-        ...linode1,
-        maintenance: null
-      }
-    ]);
-
-    expect(addNotificationsToLinodes([], [linode1])).toEqual([
-      {
-        ...linode1,
-        maintenance: null
-      }
-    ]);
-  });
+  expect(
+    addNotificationsToLinodes(maintenanceNotification(4325345345), [linode1])
+  ).toEqual([
+    {
+      ...linode1,
+      maintenance: null
+    }
+  ]);
 });
