+++ packages/manager/src/store/linodes/types.ts
@@ -1,25 +1,7 @@
-import { Linode } from 'linode-js-sdk/lib/linodes';
+import { LinodeWithMaintenance as L } from './linodes.helpers';
 
-/**
- * _when_ is not guaranteed to exist if this is a maintenance notification
- *
- * _when_ could be in the past
- *
- * In the case of maintenance, _until_ is always going to be _null_,
- * so we cannot tell the user when their maintenance window will end. :(
- */
-
-export type Type = 'reboot-scheduled' | 'migration-pending';
-
-export interface Maintenance {
-  type: Type;
-  when: string | null;
-  until: string | null;
-}
-
-export interface LinodeWithMaintenance extends Linode {
-  maintenance?: Maintenance | null;
-}
+/* tslint:disable-next-line */
+export interface LinodeWithMaintenance extends L {}
 
 export interface LinodeWithMaintenanceAndMostRecentBackup
   extends LinodeWithMaintenance {
